Refactor methods for generating alignment PDBs
Move them from biojava-structure-gui to biojava-structure,
since generating the multi-model PDB files might be generally
useful.
